creation dossier store e rename zustand.tsx en zustand.ts
